made notebook.py file
